Remove linters error
